+++ buildSrc/src/main/kotlin/Config.kt
@@ -62,7 +62,7 @@ val HOST_OS: OperatingSystem = findHostOs()
 
 object Realm {
     val ciBuild = (System.getenv("CI") != null)
-     const val version = "3.2.8"
+     const val version = "3.2.9"
     const val group = "io.github.xilinjia.krdb"
     const val projectUrl = "https://realm.io"
     const val pluginPortalId = "io.github.xilinjia.krdb"
@@ -98,12 +98,13 @@ object Realm {
 object Versions {
     object Android {
         const val minSdk = 21
-        const val targetSdk = 35
-        const val compileSdkVersion = 35
-        const val buildToolsVersion = "35.0.0"
-        const val buildTools = "8.7.3" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
+        const val targetSdk = 36
+        const val compileSdkVersion = 36
+        const val buildToolsVersion = "36.1.0"
+        const val buildTools = "8.13.0" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
         const val ndkVersion = "27.0.12077973"
-        const val r8 = "8.7.18" // See https://developer.android.com/build/kotlin-support
+//        const val ndkVersion = "29.0.14206865"    // TODO: causes lots of errors in C++ and h files
+        const val r8 = "8.10.21" // See https://developer.android.com/build/kotlin-support
     }
     const val androidxBenchmarkPlugin = "1.3.3" // https://maven.google.com/web/index.html#androidx.benchmark:androidx.benchmark.gradle.plugin
     const val androidxStartup = "1.2.0" // https://maven.google.com/web/index.html?q=startup#androidx.startup:startup-runtime
@@ -115,29 +116,29 @@ object Versions {
     const val atomicfuPlugin = "0.29.0" // https://github.com/Kotlin/kotlinx.atomicfu
     const val autoService = "1.0" // https://mvnrepository.com/artifact/com.google.auto.service/auto-service
     const val buildkonfig = "0.13.3" // https://github.com/yshrsmz/BuildKonfig
-    const val coroutines = "1.10.1" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
-    const val datetime = "0.6.1" // https://github.com/Kotlin/kotlinx-datetime
-    const val detektPlugin = "1.23.6" // https://github.com/detekt/detekt
-    const val dokka = "2.0.0" // https://github.com/Kotlin/dokka
-    const val gradlePluginPublishPlugin = "0.15.0" // https://plugins.gradle.org/plugin/com.gradle.plugin-publish
-    const val jmh = "1.34" // https://github.com/openjdk/jmh
-    const val jmhPlugin = "0.6.6" // https://github.com/melix/jmh-gradle-plugin
+    const val coroutines = "1.10.2" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
+    const val datetime = "0.7.1" // https://github.com/Kotlin/kotlinx-datetime
+    const val detektPlugin = "1.23.8" // https://github.com/detekt/detekt
+    const val dokka = "2.1.0" // https://github.com/Kotlin/dokka
+    const val gradlePluginPublishPlugin = "2.0.0" // https://plugins.gradle.org/plugin/com.gradle.plugin-publish
+    const val jmh = "1.37" // https://github.com/openjdk/jmh
+    const val jmhPlugin = "0.7.3" // https://github.com/melix/jmh-gradle-plugin
     const val junit = "4.13.2" // https://mvnrepository.com/artifact/junit/junit
     const val kbson = "0.4.0" // https://github.com/mongodb/kbson
     // When updating the Kotlin version, also remember to update /examples/min-android-sample/build.gradle.kts
     // these need to be compatible with Kotlin version: atomicfu, atomicfuPlugin, serialization, kotlinCompileTesting, ktlint
-    const val kotlin = "2.2.0" // https://github.com/JetBrains/kotlin and https://kotlinlang.org/docs/releases.html#release-details
-    const val latestKotlin = "2.2.0" // https://kotlinlang.org/docs/eap.html#build-details
+    const val kotlin = "2.2.20" // https://github.com/JetBrains/kotlin and https://kotlinlang.org/docs/releases.html#release-details
+    const val latestKotlin = "2.2.20" // https://kotlinlang.org/docs/eap.html#build-details
     const val kotlinJvmTarget = "17" // Which JVM bytecode version is kotlin compiled to.
     const val kotlinCompileTesting = "0.8.0" // https://github.com/zacsweers/kotlin-compile-testing
-    const val ktlint = "1.7.0" // https://github.com/pinterest/ktlint
+    const val ktlint = "1.7.1" // https://github.com/pinterest/ktlint
     const val ktor = "2.3.12" // https://github.com/ktorio/ktor
     const val multidex = "2.0.1" // https://developer.android.com/jetpack/androidx/releases/multidex
     const val nexusPublishPlugin = "1.3.0" // https://github.com/gradle-nexus/publish-plugin
-    const val okio = "3.10.2" // https://square.github.io/okio/#releases
+    const val okio = "3.16.2" // https://square.github.io/okio/#releases
     const val relinker = "1.4.5" // https://github.com/KeepSafe/ReLinker
-    const val serialization = "1.9.0" // https://kotlinlang.org/docs/releases.html#release-details
-    const val shadowJar =  "8.1.0" // https://mvnrepository.com/artifact/com.github.johnrengelman.shadow/com.github.johnrengelman.shadow.gradle.plugin?repo=gradle-plugins
+    const val serialization = "1.9.0" // https://github.com/Kotlin/kotlinx.sericalization
+    const val shadowJar =  "8.1.1" // https://mvnrepository.com/artifact/com.github.johnrengelman.shadow/com.github.johnrengelman.shadow.gradle.plugin?repo=gradle-plugins
     const val snakeYaml = "1.33" // https://github.com/snakeyaml/snakeyaml
     val sourceCompatibilityVersion = JavaVersion.VERSION_17 // Language level of any Java source code.
     val targetCompatibilityVersion = JavaVersion.VERSION_17 // Version of generated JVM bytecode from Java files.