+++ packages/test-base/src/commonMain/kotlin/io/github/xilinjia/krdb/test/util/Utils.kt
@@ -29,6 +29,7 @@ import kotlinx.coroutines.withTimeout
 import kotlinx.datetime.Instant
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
+import kotlin.time.ExperimentalTime
 
 // Platform independent helper methods
 object Utils {
@@ -77,6 +78,7 @@ suspend fun Realm.useInContext(action: suspend (Realm) -> Unit) {
 }
 
 // Convert Kotlinx-datatime Instant to RealmInstant
+@OptIn(ExperimentalTime::class)
 fun Instant.toRealmInstant(): RealmInstant {
     val s: Long = this.epochSeconds
     val ns: Int = this.nanosecondsOfSecond