+++ packages/test-base/src/jvmMain/kotlin/io/github/xilinjia/krdb/test/util/Compiler.kt
@@ -33,6 +33,7 @@ object Compiler {
             messageOutputStream = System.out
             @Suppress("deprecation")
             componentRegistrars = plugins
+//            compilerPluginRegistrars = plugins
             inheritClassPath = true
             kotlincArguments = listOf("-Xjvm-default=all-compatibility")
         }.compile()