+++ packages/test-sync/src/jvmMain/kotlin/io/github/xilinjia/krdb/test/mongodb/BaasUtils.kt
@@ -1,3 +1,0 @@
-package io.github.xilinjia.krdb.test.mongodb
-
-actual fun baasTestUrl(): String = SyncServerConfig.url